# ascii: rewrite function to naturally generate more optimized code