A lot of functions inside std.ascii and std.uni are fast only optimized by the
compiler optimizer. Some functions can be written to naturally generate more
optimized code.

Signed-off-by: Luís Ferreira <contact@lsferreira.net>

---

Take the example of isDigit implementation:

Here are some performance tests comparing std and new std implementations. Running with 100 000 000 randomized iterations without running the optimizer

```
DMD64 v2.095.0
2 ms, 253 μs, and 9 hnsecs
2 ms, 242 μs, and 6 hnsecs

LDC 1.24.0
2 ms, 148 μs, and 2 hnsecs
2 ms, 105 μs, and 6 hnsecs

GDC 10.2.0
2 ms, 266 μs, and 5 hnsecs
2 ms, 251 μs, and 5 hnsecs
```

NOTE: These times are for 100 000 runs because the benchmark was made with 1000 different random seeds.

Generated Assembly for both implementations without the optimizer: https://godbolt.org/z/sav5qe
Benchmark source code: https://gist.github.com/run-dlang/8198a98f4fb4e99120f6351bd764408a