@maxhaton 2021-02-17
https://godbolt.org/z/MeT148 

Look at the code generated by dmd with no optimizations - it looks longer, right? Well now consider that your processor can do *billions* of instructions per second - the only part of this code that actually makes any difference to performance in a modern processor is the branch (`cmp`) which is unavoidable in the general case. 

Also, if we plot your benchmarking code with the call to isDigitX *removed*, we will see that the tiny optimization you report is only a fraction of a standard deviation away from  the mean. 
![image](https://user-images.githubusercontent.com/5672948/108142568-f0710280-70bd-11eb-8867-fec88b4f303c.png)

The black lines are at +- (253-242)