@n8sh 2021-02-17
The benchmark doesn't make a convincing case for these changes but I find some of the code more beautiful.