@berni44 2021-02-17
We had a similar discussion at PR #7773. When I opened this PR I was convinced, that having faster code would be a good idea. After the comment from @kinke I changed my mind: Meanwhile I think, we should restrain from such optimizations if the compilers manage to do the optimizations on their own. There is little speed gain (because when one is interested in speed one will use -O and the change comes down to nothing), but the code get's worse to read and understand.