@ljmf00 2021-02-17
> [godbolt.org/z/MeT148](https://godbolt.org/z/MeT148)

Just a note, `dmd-nightly` is as old as `dmd-nighlty` in `dlang-core` from https://run.dlang.org/ . It's preferable to use the latest stable version rather than using that version. Somehow is stuck in the time. Maybe it's the same issue of DMD used in dub.

> Look at the code generated by dmd with no optimizations - it looks longer, right? Well now consider that your processor can do _billions_ of instructions per second - the only part of this code that actually makes any difference to performance in a modern processor is the branch (`cmp`) which is unavoidable in the general case.

Sure. I understand the point that optimizers will perfectly do their job, although, when we are talking about those micro-optimizations, jumps are costly, and generating code with a good CPU pipelining flow is preferable. If we take the example of a more "jumpy" generated code, `isAlpha`, we see a tremendous improvement by getting rid of 3 jumps and writing only one time to the CPU registers. And we shouldn't always focus on x86, if we look at ARM-based generated code, it's even worse!

https://godbolt.org/z/E7P4Yx

> Also, if we plot your benchmarking code with the call to isDigitX _removed_, we will see that the optimization you report is only a fraction of a standard deviation away from the mean.

Where did you get those plots btw? Looks fancy :sweat_smile: