@kinke 2021-02-17
> and generating code with a good CPU pipelining flow is preferable

If the optimizers do that and the assumption holds that people don't care about saving some nanoseconds for debug builds, it's IMO preferable to keep the code nice, simple & self-explanatory instead of having to add a comment explaining what the magic trick in the next line does.