@maxhaton 2021-02-17
I've managed to edit your comment by accident sorry. Anyway the script is a python script I use for physics stuff with the styling removed. 

This is still small fry for a function that is typically not called very often - it will be inlined