@maxhaton 2021-02-17
@ljmf00 Unfortunately I can't guesstimate the throughput of the arm code by eye and I don't have a box to test it on, but worrying about these things is really not helpful for code built with optimizations turned off - the compiler is barely doing register allocation. 

If you profile this kind of (verbose, let's say) generated code in vTune or similar, you can see that - on a microarchitectural level - the behaviour is so inconsistent that (although there is a trend) an energy spent optimizing things in this manner is rather fruitless in the long run. One problem that you may have bumped into is that doing DIY benchmarking at the level of microarchitecture can be a complete pain because even the act of extracting the data you recorded can affect the measurement.